Rename `neo4jprojection` to `opencypherprojection` package.
