+++ boltprojection/adaptor.go
@@ -4,10 +4,10 @@ import (
 	"context"
 	"time"
 
-	"github.com/deslittle/projectionkit/internal/identity"
-	"github.com/deslittle/projectionkit/internal/unboundhandler"
-	"github.com/deslittle/projectionkit/resource"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/projectionkit/internal/identity"
+	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
+	"github.com/dogmatiq/projectionkit/resource"
 	"go.etcd.io/bbolt"
 )
