+++ internal/adaptortest/suite.go
@@ -3,9 +3,9 @@ package adaptortest
 import (
 	"context"
 
-	"github.com/deslittle/projectionkit/resource"
 	"github.com/dogmatiq/dogma"
 	"github.com/dogmatiq/dogma/fixtures"
+	"github.com/dogmatiq/projectionkit/resource"
 	"github.com/onsi/ginkgo"
 	"github.com/onsi/ginkgo/extensions/table"
 	"github.com/onsi/gomega"
