+++ opencypherprojection/adaptor.go
@@ -5,10 +5,10 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/deslittle/projectionkit/internal/identity"
-	"github.com/deslittle/projectionkit/internal/unboundhandler"
-	"github.com/deslittle/projectionkit/resource"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/projectionkit/internal/identity"
+	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
+	"github.com/dogmatiq/projectionkit/resource"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
 )
