+++ opencypherprojection/resource.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/deslittle/projectionkit/resource"
+	"github.com/dogmatiq/projectionkit/resource"
 	"github.com/neo4j/neo4j-go-driver/v5/neo4j"
 )
