+++ resource/resource_test.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"errors"
 
-	. "github.com/deslittle/projectionkit/resource"
 	. "github.com/dogmatiq/dogma/fixtures"
+	. "github.com/dogmatiq/projectionkit/resource"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
