+++ sqlprojection/adaptor.go
@@ -5,10 +5,10 @@ import (
 	"database/sql"
 	"time"
 
-	"github.com/deslittle/projectionkit/internal/identity"
-	"github.com/deslittle/projectionkit/internal/unboundhandler"
-	"github.com/deslittle/projectionkit/resource"
 	"github.com/dogmatiq/dogma"
+	"github.com/dogmatiq/projectionkit/internal/identity"
+	"github.com/dogmatiq/projectionkit/internal/unboundhandler"
+	"github.com/dogmatiq/projectionkit/resource"
 )
 
 // adaptor adapts an sqlprojection.ProjectionMessageHandler to the
