+++ sqlprojection/select_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"database/sql"
 
-	. "github.com/deslittle/projectionkit/sqlprojection"
+	. "github.com/dogmatiq/projectionkit/sqlprojection"
 	"github.com/dogmatiq/sqltest/sqlstub"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
